Add AWS env check

+++ ml_enabler/api/task.py
@@ -3,6 +3,7 @@ from flask_restful import request, current_app
 from ml_enabler.utils import err
 from ml_enabler.services.task_service import TaskService
 from ml_enabler.models.utils import NotFound
+import ml_enabler.config as CONFIG
 
 task_bp = Blueprint(
     'task_bp', __name__
@@ -10,6 +11,9 @@ task_bp = Blueprint(
 
 @task_bp.route('/v1/task', methods=['GET'])
 def list():
+    if CONFIG.EnvironmentConfig.ENVIRONMENT != "aws":
+        return err(501, "stack must be in 'aws' mode to use this endpoint"), 501
+
     pred_id = request.args.get('pred_id')
     task_type = request.args.get('type')
 
@@ -29,6 +33,9 @@ def list():
 
 @task_bp.route('/v1/task/<int:task_id>', methods=['GET'])
 def get(task_id):
+    if CONFIG.EnvironmentConfig.ENVIRONMENT != "aws":
+        return err(501, "stack must be in 'aws' mode to use this endpoint"), 501
+
     try:
         return TaskService.get(task_id)
     except NotFound:
@@ -40,6 +47,9 @@ def get(task_id):
 
 @task_bp.route('/v1/task/<int:task_id>', methods=['DELETE'])
 def delete(task_id):
+    if CONFIG.EnvironmentConfig.ENVIRONMENT != "aws":
+        return err(501, "stack must be in 'aws' mode to use this endpoint"), 501
+
     try:
         return TaskService.delete(task_id)
     except NotFound:
